added cookie identification to check if you are the person who created it and fixed some minor bugs

+++ songbattle/pages/api/room/delete.ts
@@ -4,44 +4,49 @@ import { connectToDb } from "../../../utils/connectToDb";
 export default async (req: VercelRequest, res: VercelResponse) => {
   const id = req.body.id;
 
-  /**
-   * Todo: Define Request Method (if possible) and maybe add some form of authentication
-   */
-
-  if (typeof id !== "string") {
-    res.status(400).send({
-      added: false,
-      message: "Bad Input",
-    });
-    return;
-  }
-  const db = await connectToDb();
-
-  await db
-    .query("SELECT * FROM songs WHERE id=$1", [id])
-    .then((a) => {
-      if (a.rowCount <= 1) {
-        db.query("DELETE FROM room WHERE id=$1", [id])
-          .then((r) => {
-            if (a.rowCount <= 0) {
-              res.send({ deleted: false, message: "Room not found!" });
-              return;
-            }
-            res.send({ deleted: true, id: id });
-          })
-          .catch((err) =>
-            res
-              .status(400)
-              .send({ deleted: false, error: true, message: err.stack })
-          );
-        return;
-      }
+  if (req.method === "DELETE") {
+    if (typeof id !== "string") {
       res.status(400).send({
-        error: true,
-        message: "There are too many songs still in the queue!",
+        added: false,
+        message: "Bad Input",
       });
-    })
-    .catch((err) =>
-      res.status(400).send({ deleted: false, error: true, message: err.stack })
-    );
+      return;
+    }
+    const db = await connectToDb();
+
+    //check cookies
+    /*const userCookie = req.cookies.user;
+    if (userCookie === null || userCookie === "undefined")
+      return res.status(200).send({ deleted: false, error: true });*/
+
+    await db
+      .query("SELECT songlink FROM songs WHERE id=$1", [id])
+      .then((a) => {
+        if (a.rowCount <= 1) {
+          db.query("DELETE FROM room WHERE id=$1", [id])
+            .then((r) => {
+              if (a.rowCount <= 0) {
+                res.send({ deleted: false, message: "Room not found!" });
+                return;
+              }
+              res.send({ deleted: true, id: id });
+            })
+            .catch((err) =>
+              res
+                .status(400)
+                .send({ deleted: false, error: true, message: err.stack })
+            );
+          return;
+        }
+        res.status(400).send({
+          error: true,
+          message: "There are too many songs still in the queue!",
+        });
+      })
+      .catch((err) =>
+        res
+          .status(400)
+          .send({ deleted: false, error: true, message: err.stack })
+      );
+  }
 };
